+++ src/app/craft/[slug]/page.tsx
@@ -4,15 +4,11 @@ import { notFound } from 'next/navigation';
 
 import { Giscus } from '@/components/comments/giscus';
 import { SkeletonTableOfContents } from '@/components/loading/skeleton-table-of-contents';
-import {
-  createCraftMDXProcessor,
-  CustomMDX,
-  formatDate,
-} from '@/components/mdx';
 import { siteMetadata } from '@/lib/site-metadata';
+import { createCraftMDXProcessor, CustomMDX, formatDate } from '@/mdx';
 
 const TableOfContents = dynamic(
-  () => import('@/components/mdx').then(mod => mod.TableOfContents),
+  () => import('@/mdx').then(mod => mod.TableOfContents),
   {
     ssr: false,
     loading: SkeletonTableOfContents,
