--- src/components/mdx/index.ts
@@ -1,3 +1,0 @@
-export * from './lib/mdx';
-export { CustomMDX } from './ui/mdx';
-export { TableOfContents } from './ui/table-of-contents';
