+++ src/mdx/common/create-mdx-component.ts
