+++ src/mdx/common/step-content-wrapper/step-content-wrapper.tsx
