+++ src/mdx/components/img/rounded-image.tsx
