+++ src/mdx/index.ts
@@ -1,0 +1,3 @@
+export * from './mdx';
+export { CustomMDX } from './mdx';
+export { TableOfContents } from './table-of-contents';
